Completes ex. 9.1, adds calcluateBmi function
This commit adds bmiCalculator.ts which contains a function
calcluateBmi that takes two numbers as parameters, a height in cm and
a weight in kg, and returns a string with the BMI description for
those measurements. It also adds a npm script to execute
bmiCalculator.ts. Currently, parameters for calcluateBmi are
hardcoded, rather than being set in the command line.