Completes ex. 9.2, adds calcluateExercise function
This commit adds exerciseCalculator.ts which contains a function
calcluateExercise that takes two parameters - the first is an array
of numbers, representing hours of exercise per day, and the second is
a number that represents the target exercise hours per day. It
returns an object defined via an interface in exerciseCalculator.ts
which contains analysis of the exercise data submitted as parameters.
The commit also adds a npm script to execute exerciseCalculator.ts.
Currently, parameters for calcluateExercise are hardcoded, rather
than being set in the command line.