Completes ex 9.3, enables command-line variables
This commit adds functionality to both bmiCalculator.ts and
exerciseCalculator.ts to take command line variables. For the former,
two numbers are expected, the first height in cm the second weight in
kg. For the latter, at least two numbers are expected - the first is
the target, and all subsequent are daily exercise amounts. Command
line data is validated to ensure there are sufficient arguments and
that all arguments are correct types. Appropriate errors are thrown
when command line data is incorrect.