Completes ex 9.4, adds express to project
This commit adds express to the project to handle routing, as well as
relevant type definitions and ts-node-dev library. It updates
package.json with commands to start the express server in both dev
and regular modes.

The index.ts file currently has routing only for '/hello' with a
template message.

This commit also updates tsconfig.json with a great deal more
configuraitons.