Completes ex. 9.5, adds API request for BMI
This commit adds routing to '/bmi' with parameters for height and
weight that returns json of the height, weight, and bmiCalculator
function response. If the height or weight parameters are missing
from the request, or they are not numbers, instead the response is
status code 400 with an error message describing the issue.

The calculateBmi function is changed to an export, and is imported
into index.ts.