Completes ex 9.7, adds API endpoint for exercises
This commit adds an endpoint for a POST request to '/exercises'. It
takes a JSON body with a 'target' parameter that can be evaluated as
a number, and a 'daily_exercises' parameter that can be evaluated as
an array of numbers. If either parameter is missing or malformatted,
an appropriate error is sent with status code 400. Otherwise the
return object of the calculateExercise function is returned.

This commit also includes ex. 9.6, which was to set up ts linting for
the project.