Completes ex. 9.9, adds patientor FE as submodule
This commit clones the patientor frontend provided by Full Stack Open
as a submodule into the FSO_09 repo. It makes some changes to the
files around my preferences for linting, corrects the backend port to
the value I am using, and also adds CORS to the backend to ensure the
two parts communicate successfully.