Completes ex 9.11, adds route for /api/patients
This commit adds backend routing to display a list of all patients
to /api/patients. This routing excludes SSN from the data sent. The
commit creates both a type/interface for Patient which includes SSN
and creates a utility type of PatientNoPII which excludes SSN. The
data is also correctly displayed in the frontend.