Completes ex 9.12&13, adds POST for new patients
This commit adds a POST endpoint at /api/patients to add new patients
to the app. This includes parsing, validation, and type predicates to
ensure correct typing of the submitted data. This commit also creates
a Gender enum type with values of 'male', 'female', and 'non-binary'
(the frontend is modified to use 'non-binary' rather than 'other').
This endpoint works with the frontend new patient form (after it has
been modified as described above).