Completes ex 9.19, changes input types in form
This commit changes the date input in the form to an HTML date input,
and changes the visibility and weather inputs to radio buttons based
on the associated enums. The sets of radio buttons are mapped off of
the enum keys, so changes to the enum should be reflected in the page
without having to make any changes to its code. The associated state
variables are now set to type Visibility | undefined and Weather |
undefined respectively, and are set undefined on the initial render.