Completes ex 9.20, adds single patient endpoint
This commit extends the patientor backend to add an endpoint at
'/api/patients/:id' which fetches the full patient info for a single
patiend with that id. It also adds a new currently empty interface
type Entry and an entries member to the Patient interface which is an
array of Entry. The function toNewPatient in src/utils.ts is updated
to reflect the new member of Patient, as well as new associated type
narrowing functions.