Adds patientor_frontend as part of repo
This commit re-adds the frontend of the patientor app to the repo,
but not as a sub-module. This should allow me to commit changes to
the frontend, as required for further exercises in pt. 9.

This commit also reflects some frontend changes made in previous
commits that were relevant to backend setup or enhancements.