Completes ex. 9.22, adds type for BE Entry field
This commit adds additional data to the backend of patientor so that
the entry fields of the patients are now populated with information
for medical visits. The entries are of 3 types, which are implemented
as a union type between three extensions of a base entry. The router
and service for retrieving a single patient are also updated - the
service now adds validation that the ID exists, refers to only one
member of the database, and that all entries in that patient's entry
array are of one of the three types. The service now returns an
object that contains success or failure info - if success, it returns
the patient, if failure it returns an appropriate HTML code and error
message.