Completes ex. 9.27, adds HealthCheck Entry form
This commit adds a basic form to add HealthCheck entries as a part of
the PatientInfo component. The form is a new HealthCheckEntry
component, which has textboxes for all fields, a reset button, and a
submit button. If there is an issue with the data, a window.alert
error message is displayed describing the issue.