Completes ex. 9.28, adds Entry form for all types
This commit adds entry forms for Occupational Healthcare and Hopsital
Entries, to accompany the Health Check form. These are each rendered
by the new NewEntryForm component, which renders a containing box,
header, and has a dropdown selection for which type of form. The
PatientInfo component now renders the NewEntryForm component.