Completes ex. 9.29, changes Entry input elements
This commit changes the input elements in the various Entry
components to better reflect the type of data. Diagnosis codes is
changed to a multiple select with checkboxes. Heath Check Ratings is
changes to a select, with the elements to pick from being the
descriptors rather than the numerical codes. Dates are date inputs
using the MUI DatePicker element from the mui-x-datepicker library,
along with the dayjs library.